move placeholders file location
